component css modules restructured
