+++ app/layout.tsx
@@ -1,4 +1,4 @@
-import { Header } from '../components/Header'
+import { Header } from '../components/header/Header'
 import '../styles/globals.css'
 
 export default function RootLayout({ children, }: { children: React.ReactNode }) {
